+++ .gitignore
@@ -3,4 +3,4 @@ node_modules/
 /playwright-report/
 /blob-report/
 /playwright/.cache/
-./auth
+.auth/
